requirements.txt
httpx>=0.27.0
python-dotenv>=1.0.0
sentence-transformers>=2.2.2
chromadb>=0.4.24
langchain>=0.1.0
langchain-community>=0.0.16
